requirements.txt
# requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pyperclip==1.8.2
pystray==0.19.5
pillow==10.1.0
keyboard==0.13.5
requests==2.31.0
psutil==5.9.6
pywin32==306